requirements.txt
kafka-python
fastavro
pandas
jupyter